Light
Light is a collaborative mobile application that helps blind people walk around the globe. 

Main idea of the app:

People can use the app to mark places where there are problems on the streets such as holes, fallen trees, floods and etc. 
When blind people pass around these markers on the map an alert will be released informing the problem and where the it was found.